+++ endpoints/user_management/delete_account.js
@@ -1,11 +1,18 @@
 import express from 'express';
 import client from '../../db.js';
+import {verifyToken} from "../../tokenmanagement.js";
 
 const deleteRouter = express.Router();
 
 deleteRouter.delete('/:userId/delete', async (req, res) => {
-    const userId = req.params.userId;
     try {
+        const userId = req.params.userId
+
+        const decodedUserId = await verifyToken(userId);
+        if (!decodedUserId) {
+            return res.status(401).json({ error: 'Unauthorized' });
+        }
+
         await client.query('DELETE FROM users WHERE id = $1', [userId]);
         res.status(200).json({ message: 'User was successfully deleted' });
     } catch (error) {
